btrdbextras/opendss_ingest/requirements.txt
OpenDSSDirect.py[extras]

numpy>=1.24.3
tqdm>=4.66.1
